Isolate materials code
