+++ controllers/materials/helpers.js
@@ -3,7 +3,6 @@ const { reduce } = require('lodash');
 const { check } = require('express-validator/check');
 
 const { errorTranslator } = require('../../modules/validators');
-const contentApi = require('../../services/content-api');
 
 const translateError = errorTranslator('global.forms');
 const translationLabelBase = 'funding.guidance.order-free-materials.formFields.';
@@ -250,22 +249,9 @@ If you have feedback, please contact digital.monitoring@biglotteryfund.org.uk.`;
     return text.trim();
 }
 
-function injectMerchandise({ locale = false, showAll = false }) {
-    return async (req, res, next) => {
-        try {
-            const localeToUse = locale ? locale : req.i18n.getLocale();
-            res.locals.availableItems = await contentApi.getMerchandise(localeToUse, showAll);
-            next();
-        } catch (error) {
-            next(error);
-        }
-    };
-}
-
 module.exports = {
     materialFields,
     makeOrderText,
     postcodeArea,
-    injectMerchandise,
     normaliseUserInput
 };
