+++ controllers/materials/views/materials.njk
@@ -1,6 +1,6 @@
 {% from "components/hero.njk" import hero %}
 {% from "components/icons.njk" import iconClose %}
-{% from "components/forms.njk" import textInput, radios with context %}
+{% from "./materials-fields.njk" import textInput, radios with context %}
 
 {% extends "common/informationPage.njk" %}
